User creation flow properly works
Note: The redirection after user enters password, security question, and answer
has no user feedback. This may be confusing but I am unsure how to add a page.
If this is an issue, can investigate and see how to resolve. May be able to
direct via the trigger('<insert var here>', navigate) in the then after
submitting the fields.