Add error message passthrough
Currently operates by editing the errorSummary field to start with
the identifier CLQ: which is checked for by the widget. This is
done to prevent inadvertently overriding the default error message
displays. The error modification is done on the back-end in order
to prevent excessive customization to the widget application and
introducing more customized dependencies