Add fix for OKTA API recovery question bug
When an email contains a plus or a period, and possibly other
symbols, the recovery question returned is the question rather
than the question text. The implemented fix relies on a hard-coded
set of questions which the return value is checked against. If
a match is found, then the question text is returned, else,
the original return value is displayed. This is only necessary
while the bug exists in Okta's API