CancerLinQ customizations to the okta-widget
  The changes were done with the intention of being upgradeable
in order to stay updated against the official branches. However,
due to the nature of some changes, this is not guaranteed and
should be verified post-upgrade to ensure functionality remains
the same. Also, there are a number of corresponding files on
the SAP HANA systems and some of the overridden API calls are
constructed to work in coordination with those files via
an internal API endpoint structure.